This commit contains major changes to the  deterministic equivalent code and a new series of tests for understanding where inaccuracies are occuring in the deterministic equivalent code.
It updates the `theory_lambda_min` function  to make it work below alpha = 0.5.  It now requires v and d to be passed.

New methods were added to deterministic_equivalent, including a new method of generating 'deterministic_spectra' eigenvalues.  This improves accuracy at small alpha.  Strong D-dependence is observed for small alpha.  D needs to get to 3200+ for good agreement when alpha=0.3.

The method of choosing the 'split' has been modified for generating the rhos.  Also, a d-dependent split has been set as the default behavior.   deterministic_rho_weights no longer accepts a number of splits.

Additionally, several new test scripts have been added to validate the accuracy of the deterministic rho weights and the limit loss predictions against theoretical values.

It untroduces a new empirical limit loss method in the `PowerLawRF` class, for comparison against the theory fit.

The usage of 'RISK' has been standardized to mean MSE/2.  The ODE code, which follows the convention in the paper outputs twice-the-risk.  The documentation (and variable nomenclature) has been written to correctly reflect this.

See det-equivalent-low-alpha-test.py for the suggested method of running the ODEs.